// apps/web/src/proxy/models/roleResponse.ts
// AUTO-GENERATED — DO NOT EDIT.
// Regenerate with: npm run generate:proxy
// Source: StarterKit.WebApi | v1 1.0.0

export interface RoleResponse {
  id?: string;
  name?: string;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
  isElevated?: boolean;
  isPortalRole?: boolean;
  isSystem?: boolean;
  /** @nullable */
  clubId?: string | null;
  isDefault?: boolean;
  permissions?: string[];
}